requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
pillow
sqlalchemy